Deleted test class CompletionServiceDefaultTest2
