+++ apps/onboarding-functions/src/test/java/it/pagopa/selfcare/onboarding/service/CompletionServiceDefaultTest.java
@@ -23,7 +23,9 @@ import jakarta.ws.rs.core.Response;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
 import org.jboss.resteasy.core.ServerResponse;
 import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mockito;
 import org.openapi.quarkus.core_json.api.DelegationApi;
@@ -147,7 +149,7 @@ public class CompletionServiceDefaultTest {
         InstitutionResponse serviceResponse = completionServiceDefault.createOrRetrieveInstitution(onboarding);
 
         assertNotNull(serviceResponse);
-        assertEquals(serviceResponse.getId(), "actual-id");
+        assertEquals("actual-id", serviceResponse.getId());
     }
 
     @Test
@@ -716,6 +718,36 @@ public class CompletionServiceDefaultTest {
 
         Mockito.verify(notificationService, times(1))
                 .sendTestEmail(executionContext);
+    }
+
+    @Nested
+    @TestProfile(CompletionServiceDefaultTest.ForceCreationProfile.class)
+    @TestInstance(TestInstance.Lifecycle.PER_CLASS)
+    class ForceCreationProfile implements QuarkusTestProfile {
+
+        @Override
+        public Map<String, String> getConfigOverrides() {
+            return Map.of("onboarding-functions.force-institution-persist", "true");
+        }
+
+        @Test
+        void forceInstitutionCreationFlagTrue(){
+            Onboarding onboarding = createOnboarding();
+
+            Institution institutionSa = new Institution();
+            institutionSa.setTaxCode("taxCode");
+            institutionSa.setInstitutionType(InstitutionType.SA);
+            institutionSa.setOrigin(Origin.ANAC);
+            onboarding.setInstitution(institutionSa);
+
+            InstitutionResponse institutionResponse = dummyInstitutionResponse();
+            when(institutionApi.createInstitutionFromAnacUsingPOST(any())).thenReturn(institutionResponse);
+
+            completionServiceDefault.createInstitutionAndPersistInstitutionId(onboarding);
+
+            verify(institutionApi, times(1)).createInstitutionFromAnacUsingPOST(any());
+        }
+
     }
 
     private User createDummyUser(Onboarding onboarding) {
