--- apps/onboarding-functions/src/test/java/it/pagopa/selfcare/onboarding/service/CompletionServiceDefaultTest2.java
@@ -1,81 +1,0 @@
-package it.pagopa.selfcare.onboarding.service;
-
-import io.quarkus.test.InjectMock;
-import io.quarkus.test.junit.QuarkusTest;
-import io.quarkus.test.junit.QuarkusTestProfile;
-import io.quarkus.test.junit.TestProfile;
-import it.pagopa.selfcare.onboarding.common.InstitutionType;
-import it.pagopa.selfcare.onboarding.common.Origin;
-import it.pagopa.selfcare.onboarding.entity.Billing;
-import it.pagopa.selfcare.onboarding.entity.Institution;
-import it.pagopa.selfcare.onboarding.entity.Onboarding;
-import jakarta.inject.Inject;
-import org.eclipse.microprofile.rest.client.inject.RestClient;
-import org.junit.jupiter.api.Test;
-import org.openapi.quarkus.core_json.api.InstitutionApi;
-import org.openapi.quarkus.core_json.model.InstitutionResponse;
-
-import java.util.List;
-import java.util.Map;
-
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
-
-@QuarkusTest
-@TestProfile(CompletionServiceDefaultTest2.ForceCreationProfile.class)
-public class CompletionServiceDefaultTest2 {
-    final String productId = "productId";
-    @Inject
-    CompletionServiceDefault completionServiceDefault;
-
-    @RestClient
-    @InjectMock
-    InstitutionApi institutionApi;
-    public static class ForceCreationProfile implements QuarkusTestProfile {
-        @Override
-        public Map<String, String> getConfigOverrides() {
-            return Map.of("onboarding-functions.force-institution-persist", "true");
-        }
-    }
-
-
-    @Test
-    void forceInstitutionCreationFlagTrue(){
-        Onboarding onboarding = createOnboarding();
-
-        Institution institutionSa = new Institution();
-        institutionSa.setTaxCode("taxCode");
-        institutionSa.setInstitutionType(InstitutionType.SA);
-        institutionSa.setOrigin(Origin.ANAC);
-        onboarding.setInstitution(institutionSa);
-
-        InstitutionResponse institutionResponse = dummyInstitutionResponse();
-        when(institutionApi.createInstitutionFromAnacUsingPOST(any())).thenReturn(institutionResponse);
-
-        completionServiceDefault.createInstitutionAndPersistInstitutionId(onboarding);
-
-        verify(institutionApi, times(1)).createInstitutionFromAnacUsingPOST(any());
-    }
-
-    private Onboarding createOnboarding() {
-        Onboarding onboarding = new Onboarding();
-        onboarding.setId(onboarding.getId());
-        onboarding.setProductId(productId);
-        onboarding.setPricingPlan("pricingPlan");
-        onboarding.setUsers(List.of());
-        onboarding.setInstitution(new Institution());
-        onboarding.setUserRequestUid("example-uid");
-
-        Billing billing = new Billing();
-        billing.setPublicServices(true);
-        billing.setRecipientCode("example");
-        billing.setVatNumber("example");
-        onboarding.setBilling(billing);
-        return onboarding;
-    }
-    private InstitutionResponse dummyInstitutionResponse() {
-        InstitutionResponse response = new InstitutionResponse();
-        response.setId("response-id");
-        return  response;
-    }
-}
